# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: LibreRun
  version: 1.1.0-beta.1
paths:
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/google:
    post:
      tags:
      - auth
      summary: Login Google
      operationId: login_google_api_v1_auth_google_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/microsoft:
    post:
      tags:
      - auth
      summary: Login Microsoft
      operationId: login_microsoft_api_v1_auth_microsoft_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MicrosoftLoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/logout:
    post:
      tags:
      - auth
      summary: Logout
      operationId: logout_api_v1_auth_logout_post
      responses:
        '204':
          description: Successful Response
      security:
      - HTTPBearer: []
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    GoogleLoginRequest:
      properties:
        credential:
          type: string
          title: Credential
      type: object
      required:
      - credential
      title: GoogleLoginRequest
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MicrosoftLoginRequest:
      properties:
        code:
          type: string
          title: Code
        redirect_uri:
          anyOf:
          - type: string
          - type: 'null'
          title: Redirect Uri
      type: object
      required:
      - code
      title: MicrosoftLoginRequest
...
    TokenResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      title: TokenResponse
...
    UserProfile:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        is_platform_admin:
          type: boolean
          title: Is Platform Admin
          description: 'Whether this user administers the deployment: an admin of
            the tenant whose slug is PLATFORM_TENANT_SLUG. A display hint for the
            UI; every route keeps its own gate.'
      type: object
      required:
      - id
      - email
      - role
      - tenant_id
      - is_platform_admin
      title: UserProfile
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer